requirements.txt
astroid==2.2.5
entrypoints==0.3
flake8==3.7.7
isort==4.3.21
lazy-object-proxy==1.4.1
mccabe==0.6.1
pycodestyle==2.5.0
pyflakes==2.1.1
pylint==2.3.1
six==1.12.0
typed-ast==1.4.0
wrapt==1.11.2

# Django
Django==2.2.3
djangorestframework==3.9.4
